pusher-client-python/requirements.txt
pusher==3.3.3
websocket-client==1.8.0
